Move admin functions from views to viewsAdmin
